requirements.txt
# Core framework dependencies
langgraph>=0.0.40
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM providers
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0

# HTTP client and web framework
httpx>=0.25.0
flask>=2.3.0

# Utility libraries
jinja2>=3.1.0
uuid>=1.30

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional dependencies for enhanced functionality
aiofiles>=23.0.0
asyncio-mqtt>=0.13.0
websockets>=11.0.0